:recycle: Merge ProcessModelExecutionAdapter into ProcessModelService
