+++ src/adapters/index.ts
@@ -1,2 +1,1 @@
 export * from './notification_adapter';
-export * from './process_model_execution_adapter';
